official alpha 1

+++ src/controller/GuestUserDashboardController.java
@@ -3,6 +3,8 @@ package controller;
 import dbservice.UserDAO;
 import dbservice.UserDAODB;
 import javafx.application.Platform;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.stage.Stage;
 import model_rework.*;
 import view.DashboardView;
@@ -11,6 +13,7 @@ import view.View;
 import java.io.File;
 import java.sql.Connection;
 import java.util.ArrayList;
+import java.util.Optional;
 
 public class GuestUserDashboardController extends DashboardController {
 
@@ -24,9 +27,27 @@ public class GuestUserDashboardController extends DashboardController {
 
 		this.primaryStage = primaryStage;
 		this.primaryStage.setOnCloseRequest(e -> {
-			UserDAO UD = new UserDAODB(connection);
-			UD.deleteUser(profilemodel.getUser().getUser_id());
-			Platform.exit();
+
+			Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+			alert.setTitle("Are you sure?");
+			alert.setHeaderText("Do you want to register first?");
+			alert.setContentText("Not registering an account means all guest user info will be deleted.");
+
+			ButtonType yes = new ButtonType("Yes");
+			ButtonType no = new ButtonType("No");
+
+			alert.getButtonTypes().setAll(yes, no);
+
+			Optional<ButtonType> result = alert.showAndWait();
+			if (result.get() == yes){
+				e.consume();
+				UserRegisterController register = new GuestUserRegisterController(profilemodel.getUser().getUser_id(), profileStage, playerStage, uploadStage, primaryStage, connection, player);
+			} else {
+				UserDAO UD = new UserDAODB(connection);
+				UD.deleteUser(profilemodel.getUser().getUser_id());
+				Platform.exit();
+			}
+
 		});
 		playerStage = new Stage();
 		profileStage = new Stage();
@@ -68,13 +89,29 @@ public class GuestUserDashboardController extends DashboardController {
 	}
 
 	public void logout(){
-		UserDAO UD = new UserDAODB(connection);
-		UD.deleteUser(profilemodel.getUser().getUser_id());
-		playerStage.close();
-		uploadStage.close();
-		profileStage.close();
-		player.endPlayer();
-		LoginController login = new LoginController(primaryStage);
+
+		Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+		alert.setTitle("Are you sure?");
+		alert.setHeaderText("Do you want to register first?");
+		alert.setContentText("Not registering an account means all guest user info will be deleted.");
+
+		ButtonType yes = new ButtonType("Yes");
+		ButtonType no = new ButtonType("No");
+
+		alert.getButtonTypes().setAll(yes, no);
+
+		Optional<ButtonType> result = alert.showAndWait();
+		if (result.get() == yes){
+			UserRegisterController register = new GuestUserRegisterController(profilemodel.getUser().getUser_id(), profileStage, playerStage, uploadStage, primaryStage, connection, player);
+		} else {
+			UserDAO UD = new UserDAODB(connection);
+			UD.deleteUser(profilemodel.getUser().getUser_id());
+			playerStage.close();
+			uploadStage.close();
+			profileStage.close();
+			player.endPlayer();
+			LoginController login = new LoginController(primaryStage);
+		}
 	}
 
 	@Override
